overhaul of file organization
