+++ app/creators/universe.py
@@ -1,139 +1,14 @@
-import yaml
+
 from numpy import random as r
-import os
-
-from . import maths
-from . import language
 
 
-pdata = yaml.safe_load(open(os.path.join(os.getenv("abspath"),"app/configurations/planet.yaml")))["planet_types"]
-mdata = yaml.safe_load(open(os.path.join(os.getenv("abspath"),"app/configurations/moon.yaml")))["moon_types"]
-sdata = yaml.safe_load(open(os.path.join(os.getenv("abspath"),"app/configurations/star.yaml")))
+from ..functions import maths
+from ..functions import language
+from ..functions import configurations
 
-class Resource:
-    def __init__(self, conf):
-        self.objid = maths.uuid(n=13)
-        self.label = "resource"
-        self.conf = conf
-        self.name = conf["name"]
-        self.description = conf["description"]
-        self.volume = maths.rnd(conf["mean"],conf["std"], min_val=0)
-        if conf.get('replenish_rate'):
-            self.replenish_rate = conf['replenish_rate']
-        else: 
-            self.replenish_rate = None
-    def get_data(self):
-        fund = {
-            "name": self.name,
-            "objid": self.objid,
-            "label": self.label,
-            "volume": self.volume,
-            "description": self.description
-        }
-        if (self.replenish_rate):
-            fund['replenish_rate'] = self.replenish_rate
-        return fund
+from objects import celestials
 
-    def __repr__(self) -> str:
-        return f"<{self.label}: {self.objid}; {self.name}>"  
-
-class Body:
-    def __init__(self):
-        self.objid = maths.uuid(n=13)
-        self.type = "celestial body"
-        self.label = "body"
-        self.name = "unnamed"
-        self.resources = []
-
-    def make_name(self, n1, n2):
-        self.name = language.make_word(maths.rnd(n1, n2))
-
-    def get_fundimentals(self):
-        return {
-            "name": self.name,
-            "class": self.type,
-            "objid": self.objid,
-            "label": self.label,
-        }
-    
-    def scan_body(self):
-        for n in pdata[self.type]['resources'].keys():
-            self.resources.append(Resource(pdata[self.type]['resources'][n]))
-
-    def __repr__(self) -> str:
-        return f"<{self.label}: {self.type}; {self.objid}; {self.name}>"
-
-
-class Star(Body):
-    def build_attr(self, sdata):
-        self.make_name(1, 1)
-        self.label = "star"
-        self.type = sdata["class"]
-        self.radius = sdata["radius"]
-
-    def get_data(self):
-        fund = self.get_fundimentals()
-        fund['radius'] = self.radius
-        fund['class'] = self.type
-        return fund
-
-
-class Planet(Body):
-    def build_attr(self, t, orbiting):
-        self.make_name(2, 1)
-        self.label = "planet"
-        self.type = t
-        self.radius = maths.rnd(pdata[t]["radius_mean"], pdata[t]["radius_std"], min_val=0,type='float')
-        self.mass = maths.rnd(pdata[t]["mass_mean"], pdata[t]["mass_std"],min_val=0,type='float')
-        self.orbitsDistance = maths.np.round(maths.np.random.uniform(pdata[t]["distance_min"], pdata[t]["distance_max"]),3)
-        self.orbitsId = orbiting["objid"]
-        self.orbitsName = orbiting["name"]
-        self.isSupportsLife = False
-        self.isPopulated = False
-        self.isSurveyed = False
-
-
-    def get_data(self):
-        fund = self.get_fundimentals()
-        fund["radius"] = self.radius
-        fund["mass"] = self.mass
-        fund["orbitsDistance"] = self.orbitsDistance
-        fund["orbitsId"] = self.orbitsId
-        fund["orbitsName"] = self.orbitsName
-        fund["isSupportsLife"] = self.isSupportsLife
-        fund["isPopulated"] = self.isPopulated
-        fund["type"] = self.type
-        return fund
-
-
-class Moon(Body):
-    def build_attr(self, t, planets):
-        self.make_name(2, 1)
-        self.label = "moon"
-        self.type = t
-        self.orbiting = r.choice(planets)
-        self.orbitsId = self.orbiting["objid"]
-        self.orbitsDistance = maths.rnd(0.005,0.1,type='float')  
-        self.mass = abs(r.normal(mdata[t]["mass_mean"], mdata[t]["mass_std"]))
-        self.radius = (
-            abs(r.normal(mdata[t]["radius_mean"], mdata[t]["radius_std"]))
-            * self.orbiting["radius"]
-        )
-        self.orbitsName = self.orbiting["name"]
-        self.isSupportsLife = False
-        self.isPopulated = False
-
-    def get_data(self):
-        fund = self.get_fundimentals()
-        fund["orbitsId"] = self.orbitsId
-        fund["orbitsName"] = self.orbitsName
-        fund["orbitsDistance"] = self.orbitsDistance + self.orbiting['radius']
-        fund["mass"] = self.mass
-        fund["radius"] = self.radius
-        fund["isSupportsLife"] = self.isSupportsLife
-        fund["isPopulated"] = self.isPopulated
-        fund["class"] = self.type
-        return fund
+conf = configurations.get_configurations()
 
 
 def make_system():
@@ -143,7 +18,7 @@ def make_system():
         "label": "system",
         "isHomeSystem":"true",
         "objid": systemid,
-        "glat": maths.np.round(r.normal(0,20),3), # roughly based on NASA Exoplanet archive
+        "glat": maths.np.round(r.normal(0,20),3),
         "glon": maths.np.round(r.normal(0,20),3),
         "gelat": maths.np.round(r.normal(0,5),3)
     }
@@ -151,19 +26,19 @@ def make_system():
 
 
 def make_star():
-    s = Star()
-    s.build_attr(sdata)
+    s = celestials.Star()
+    s.build_attr(conf['star_config'])
     return s.get_data()
 
 
 def make_planet(t, orbiting):
-    p = Planet()
+    p = celestials.Planet()
     p.build_attr(t, orbiting)
     return p.get_data()
 
 
 def make_homeworld(orbiting, data):
-    p = Planet()
+    p = celestials.Planet()
     p.build_attr("terrestrial", orbiting)
     # p.scan_body()
     planet = p.get_data()
@@ -181,7 +56,7 @@ def make_homeworld(orbiting, data):
 
 
 def make_moon(t, planets):
-    m = Moon()
+    m = celestials.Moon()
     m.build_attr(t, planets)
     return m.get_data()
 
@@ -192,7 +67,7 @@ def build_homeSystem(data, username):
     star = make_star()
     planets = [
         make_planet(
-            r.choice(list(pdata.keys()), p=[pdata[t]["prob"] for t in pdata.keys()]),
+            r.choice(list(conf['planet_config'].keys()), p=[conf['planet_config'][t]["prob"] for t in conf['planet_config'].keys()]),
             star,
         )
         for p in range(int(data["num_planets"]) - 1)
@@ -201,7 +76,7 @@ def build_homeSystem(data, username):
     planets.append(homeworld)
     moons = [
         make_moon(
-            r.choice(list(mdata.keys()), p=[mdata[t]["prob"] for t in mdata.keys()]),
+            r.choice(list(conf['moon_config'].keys()), p=[conf['moon_config'][t]["prob"] for t in conf['moon_config'].keys()]),
             planets,
         )
         for p in range(int(data["num_moons"]))
